requirements.txt
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
certifi==2023.11.17
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
distro==1.8.0
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.104.1
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
marshmallow==3.20.1
motor==3.3.2
numpy==1.26.2
openai==1.3.4
orjson==3.9.10
packaging==23.2
pandas==2.1.3
pandas-stubs==2.1.1.230928
passlib==1.7.4
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.1
pydantic-extra-types==2.1.0
pydantic-settings==2.1.0
pydantic_core==2.14.3
pymongo==4.6.0
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tqdm==4.66.1
types-pytz==2023.3.1.1
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
umongo==3.1.0
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==12.0

reportlab~=4.0.7